third_party/WebKit/LayoutTests/http/tests/devtools/console/console-log-eval-syntax-error.html

<html>
<head>
<script src="../../inspector/inspector-test.js"></script>
<script src="../../inspector/console-test.js"></script>
<script>
function foo()
{
    eval("}");
}
function boo()
{
    eval("\n}\n//# sourceURL=boo.js");
}
</script>
<script>
function test() {
  TestRunner.evaluateInPage('setTimeout(foo, 0);', ConsoleTestRunner.waitUntilMessageReceived.bind(this, step2));

  function step2() {
    TestRunner.evaluateInPage('setTimeout(boo, 0);', ConsoleTestRunner.waitUntilMessageReceived.bind(this, step3));
  }

  function step3() {
    ConsoleTestRunner.dumpConsoleMessages();
    TestRunner.completeTest();
  }
}
</script>
</head>
<body onload="runTest()">
<p>
Tests that syntax errors in eval are logged into console, contains correct link and doesn't cause browser crash.
</p>
</body>
</html>
