third_party/WebKit/LayoutTests/http/tests/devtools/console/console-log-eval-syntax-error-expected.txt

CONSOLE ERROR: line 1: Uncaught SyntaxError: Unexpected token }
CONSOLE ERROR: line 2: Uncaught SyntaxError: Unexpected token }
Tests that syntax errors in eval are logged into console, contains correct link and doesn't cause browser crash.

VM:1 Uncaught SyntaxError: Unexpected token }
    at foo (console-log-eval-syntax-error.html:8)
foo @ console-log-eval-syntax-error.html:8
setTimeout (async)
(anonymous) @ VM:1
boo.js:2 Uncaught SyntaxError: Unexpected token }
    at boo (console-log-eval-syntax-error.html:12)
boo @ console-log-eval-syntax-error.html:12
setTimeout (async)
(anonymous) @ VM:1
